+++ src/index.js
@@ -1,8 +1,8 @@
-import UserInputModel from "./user-input/model.js";
-import UserInputView from "./user-input/view.js";
+import UserInputModel from "./user-input/userInputModel.js";
+import UserInputView from "./user-input/userInputView.js";
 
-import GameRuleModel from "./game-manager/model.js";
-import GameRuleView from "./game-manager/view.js";
+import GameRuleModel from "./game-manager/gameManagerModel.js";
+import GameRuleView from "./game-manager/gameManagerView.js";
 
 import { RESULT_CODE } from "./constants.js";
